a store demo

+++ src/main/java/com/lar/store/service/ProductService.java
@@ -5,6 +5,9 @@ import com.lar.store.pojo.Category;
 import com.lar.store.pojo.Product;
 import com.lar.store.util.Page4Navigator;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.cache.annotation.CacheConfig;
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -13,14 +16,25 @@ import org.springframework.stereotype.Service;
 import org.springframework.util.ClassUtils;
 
 import java.io.File;
+import java.util.List;
 
 @Service
+@CacheConfig(cacheNames="products")
 public class ProductService {
     @Autowired
     ProductDAO productDAO;
     @Autowired
     CategoryService categoryService;
+    //获取类别下的所有商品
+//    @Cacheable(key="'products-cid-'+ #p0.id")
+    public List<Product> getAllProductByCid(int cid){
+        List<Product> allByCategoryOrderByIdDesc = productDAO.findAllById(cid);
+        System.out.println(allByCategoryOrderByIdDesc);
+        return  allByCategoryOrderByIdDesc;
+
+    }
     //查询商品
+    @Cacheable(key="'products-cid-'+#p0+'-page-'+#p1 + '-' + #p2 ")
     public Page4Navigator<Product> list(int cid, int start, int size, int navigatePages) {
         Category category = categoryService.get(cid);
         Pageable pageable =PageRequest.of(start,size,Sort.Direction.DESC, "id");
@@ -28,6 +42,7 @@ public class ProductService {
         return new Page4Navigator<>(pageFromJPA,navigatePages);
     }
     //添加商品
+//    @CacheEvict(allEntries=true)
     public void add(Product bean) {
         productDAO.save(bean);
     }
@@ -36,11 +51,13 @@ public class ProductService {
         return productDAO.getOne(id);
     }
     //删除商品
+//    @CacheEvict(allEntries=true)
     public void deleteBean(int id){
         productDAO.deleteById(id);
         String uri= ClassUtils.getDefaultClassLoader().getResource("").getPath()+"static/img/product";
         File imageFolder= new File(uri);
         File file = new File(imageFolder,id+".jpg");
+        if (file!=null)
         file.delete();
     }
